+++ resources/analytics.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# require curl
+apt-get -y -q install curl >/dev/null
+
 . /vagrant/resources/colors.sh
 
 ANALYTICS=UA-66226242-7
@@ -97,7 +100,6 @@ QUERY="${QUERY} --data-urlencode \"z=${EPOCH}\""
 QUERY="${QUERY} --user-agent \"${USER_AGENT}\""
 QUERY="${QUERY} --compressed --silent --output /dev/null"
 QUERY="\"http://www.google-analytics.com/collect\" ${QUERY}"
-export toto=$QUERY
 
 if [ ! -z $HELP ]; then
 	display_help