+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## v1.2.11 (2016-07-13)
+* improved build script
+* updated nginx-rtmp-module to version 1.1.8
+* updated NodeJS to version 6.3.0
+* updated NPM to version 3.10.5
+
 ## v1.2.10 (2016-07-04)
 * fixed Debian reference box to version 8.5.0 as 8.5.1 is not usable as is
-* updated VirtualBox Gust additions to version 5.0.24
+* updated VirtualBox Guest additions to version 5.0.24
 * updated NPM to version 3.10.3
 * updated Newman to version 2.1.2